Update Install-PSD.ps1
Alright, made some massive changes to the way the script is structured. It may still be broken, I'm still working on it.

Had a version that had most of this stuff changed and amended, but somehow lost it this morning like an idiot ... it's ok though. Gonna be integrating Hyper-V shortly.