Update Install-PSD.ps1
Noticed a small error on the part I reworked a bunch of times.

Will be adding an uninstall function if the versions of ADK/WinPE/MDT aren't up to snuff. This just downloads the latest version if it misses the flags, it doesn't yet properly uninstall the previous versions. I was working on that, but something tells me doing an application removal through WMI or CIM is not the entirety of app removal... so. workin on it.