Create Provision-Hybrid.ps1
Provisions Hybrid. Had a lot of overlapping ideas, decided to put it all into one contiguous script. Still sorting out function calls in the most efficient way. It is what it is I guess. ¯\_(ツ)_/¯

I've got another module that loads a GUI and tests a credential against Active Directory or a workgroup's (SPWP|SPWI::GC()).IsInRole('nameofcoolpeople')... and if you ain't an admin? you ain't runnin' the script. But. It's not workin the way I want it to. So I left it out.

Will have more updates later today.