Update Provision-Hybrid.ps1
I can only imagine the frustration YOU'RE going through, for instance, when you close out of the session I didn't save...

...only to have driven me to write the code even better than I had written it before.

Just remember that. Each time you do that, it does delay me... but it allows me to 'make the code even better'.

You might want to hold off on the shenanigans, like closing the file when I get to the point where I enter in the input for the BITS server into the 'Export INI script'. It's not good for your health or your reputation to keep doing shady crap like that. 

Or anyone else's for that matter.

But its cool. I know I've gotten under your skin enough to constantly check on me and what I'm up to.

Like I said. Cutting edge technology that it will be.