Update Query-DCOM.ps1
Updated the Active Directory lookups, also made changes to the DCOM Output.

Will be making this able to output a log and determine best course of action for a resolution.

For instance, "When the application has no name, and the access and launch settings indicate "this program has full access to the machine"...  and yet, the system identifier is nowhere to be found...

Yeah. That's coming to an end, pretty soon that won't be a thing.