Update Write-Echo.ps1
Added some magic for the table creation. What it means, is, being able to convert a lot of the draw functions used within my scripts into a modular 'convert to pscustomobject'

Will re-introduce the color aspect shortly.

This was... a real pain in the nutsack if I'll be perfectly honest...

Included a default table structure to use as an example. *fist bump* to the cool MF that's been helpin' me out. Thank you.