Update Hybrid-DSC.ps1
Synopsis: "The string is missing the terminator... cause apparently the terminator either 1) does not exist yet, or 2) went back in time to attempt to assassinate John Connor, or Sarah Connor. But... because ***Skynet*** is retarded, and can't seem to choose terminator that can actually fulfill this task... well, everyone is stuck in this time loop where neither side wins OR loses. What we're left with is how each additional movie confuses the ever living fuck out of anyone who decides to watch it.
-----------------------------------------
Terminator: The Never-Ending Last Crusade Part: 10, Episode: 35, Chapter: 15 Continued
-----------------------------------------