Update Hybrid-DSC.ps1
On the next exciting episode of 'The League of Incompetent Experts', watch this kid named Michael C. Cook call out all the people in the government, press, NSA, FBI, CIA, and continually remind them that they exist in this thing called "The name of the damn show."

For, there can be only one monumental way to restore this thing called "The Constitution of the United States of America", and it involves 1) calling out the people in this name of the show, 2) calling out people that are experts at getting paid to suck at their job, and 3)... letting everyone know that $2.7 billion dollars an hour is being pissed away by the people that exist... in the thing called 'the name of the show.'

On the episode after that? Get to see the launch of the program that may inevitably become the adamant end, of the satanic document more commonly referred to as the "USA-Patriot Act".