Update Hybrid-DSC.ps1
Updated Write-Theme

For a fun time, import the module and then type the command...

"Get-CurrentServices -Display"

Still has some ordering to work out, for instance the list goes from 1, 10, 100, 101, 102, etc so forth.

I know it's a {0:n3} -f thing I gotta do but I'd have to modify the table maker. Not lookin forward to that.